+++ src/index.ts
@@ -200,8 +200,8 @@ interface RemoveAction extends Action {
  * @param opts - The optional configuration options.
  * @returns A new instance of the {@linkcode Tiged} class.
  */
-function tiged(src: string, opts?: Options) {
-	return new Tiged(src, opts);
+export function degit(src: string, opts?: Options) {
+	return new Degit(src, opts);
 }
 
 /**
@@ -1039,9 +1039,3 @@ async function updateCache(
 		JSON.stringify(cached, null, '  ')
 	);
 }
-
-declare namespace tiged {
-	export type { Options, Repo };
-}
-
-export default tiged;