+++ test/index.test.ts
@@ -3,7 +3,7 @@ import fs from 'node:fs/promises';
 import path from 'node:path';
 import { promisify } from 'node:util';
 import { rimraf } from 'rimraf';
-import tiged from 'tiged';
+import { degit } from 'tiged';
 
 const exec = promisify(child_process.exec);
 const tigedPath = process.env.CI