+++ tsup.config.ts
@@ -1,4 +1,3 @@
-import fs from 'node:fs/promises';
 import path from 'node:path';
 import type { Options } from 'tsup';
 import { defineConfig } from 'tsup';
@@ -23,16 +22,8 @@ export default defineConfig(options => {
 		{
 			...commonOptions,
 			format: ['cjs'],
-			dts: { footer: 'export = tiged' },
-			entry: ['src/index.ts'],
-			// https://github.com/egoist/tsup/issues/572
-			footer: {
-				js: `if (module.exports.default) {
-  Object.assign(module.exports.default, module.exports);
-  module.exports = module.exports.default;
-  delete module.exports.default;
-}`
-			}
+			dts: true,
+			entry: ['src/index.ts']
 		},
 		{
 			...commonOptions,
@@ -42,24 +33,4 @@ export default defineConfig(options => {
 			minify: true
 		}
 	];
-});
-
-// https://github.com/egoist/tsup/issues/700
-process.on('beforeExit', async code => {
-	if (code === 0) {
-		const filePath = path.resolve('dist/index.d.ts');
-		try {
-			await fs.access(filePath);
-			const file = await fs.readFile(filePath, 'utf-8');
-			const lines = file.split('\n');
-			const newContent = lines
-				.filter(line => !line.startsWith('export {'))
-				.join('\n');
-			await fs.writeFile(filePath, newContent);
-			process.exit(0);
-		} catch (err) {
-			console.error(err);
-			process.exit(1);
-		}
-	}
 });