+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Renames degit references to tiged ([#102](https://github.com/tiged/tiged/pull/102))
 
+- Breaking changes:
+
+  - Change default export to a named export ([#101](https://github.com/tiged/tiged/pull/101))
+
 - Migrate codebase to TypeScript ([#89](https://github.com/tiged/tiged/pull/89)):
   - Replace mocha with Vitest as it has better ESM and TypeScript support.
   - Use `tsup` to output both CJS and ESM entry points.